--- common/types/head_trackable.go
@@ -1,14 +1,0 @@
-package types
-
-import (
-	"context"
-)
-
-// HeadTrackable is implemented by the core txm,
-// to be able to receive head events from any chain.
-// Chain implementations should notify head events to the core txm via this interface.
-//
-//go:generate mockery --quiet --name HeadTrackable --output ./mocks/ --case=underscore
-type HeadTrackable[H Head[BLOCK_HASH], BLOCK_HASH Hashable] interface {
-	OnNewLongestChain(ctx context.Context, head H)
-}
